build: enable signed vs unsigned warnings on gcc
This enables selected things from -Wextra, can't use -Wextra because it is
fussy enough to complain about unused params on functions... they are
there for a reason.

-Wsign-compare
-Wignored-qualifiers
not -Wimplicit-fallthrough=3 ... only on gcc 7
-Wtype-limits
-Wuninitialized
not -Wclobbered ... only on gcc 7ish

fix the warnings everywhere they were found.

+++ lib/jws/jws.c
@@ -256,9 +256,9 @@ lws_jws_confirm_sig(const char *in, size_t len, struct lws_jwk *jwk)
 
 	/* 2) find length of first, hdr, block */
 
-	while (in[pos] != '.' && pos < len)
+	while (in[pos] != '.' && pos < (int)len)
 		pos++;
-	if (pos == len)
+	if (pos == (int)len)
 		return -1;
 
 	/* 3) Decode the header block */
